move tests from qa folder to tests folder and build with CMake
