+++ test/paqa_errs.c
