+++ test/qa_tools.h
